refactor: 카드 삭제 함수를 ItemButtonContainer의 역할로 분리
