+++ components/ColumnItem/ui/createItemButtonContainer.js
@@ -5,12 +5,49 @@ import {
   createElement,
   createImg,
 } from "../../../dom.js";
+import {
+  loadLocalStorage,
+  saveLocalStorage,
+} from "../../../utils/localStorage.js";
 import createEditModal from "../../Modal/createEditModal.js";
 import createModal from "../../Modal/createModal.js";
 
 const createItemButtonContainer = ({ sectionId, itemId }) => {
+  const deleteCard = ({ sectionId, itemId }) => {
+    const $columnItem = document.querySelector(
+      `#${sectionId} .column__item[data-id="${itemId}"]`
+    );
+
+    const todoList = loadLocalStorage();
+
+    const filteredList = todoList.map((section) =>
+      section.id === sectionId
+        ? {
+            ...section,
+            items: section.items.filter((item) => item.id !== itemId),
+          }
+        : section
+    );
+
+    const itemLength = filteredList.find((section) => section.id === sectionId)
+      .items.length;
+
+    const $columnCount = $columnItem
+      .closest(".column__container")
+      .querySelector(".column__count");
+
+    $columnCount.textContent = itemLength;
+
+    saveLocalStorage(filteredList);
+
+    $columnItem.remove();
+  };
+
   const handleClickDelete = () => {
-    createModal({ content: "선택한 카드를 삭제할까요?", sectionId, itemId });
+    createModal({
+      content: "선택한 카드를 삭제할까요?",
+      onClick: () => deleteCard({ sectionId, itemId }),
+    });
   };
 
   const handleClickEdit = () => {
@@ -56,7 +93,6 @@ const createItemButtonContainer = ({ sectionId, itemId }) => {
   $editButton.appendChild($editImg);
 
   $itemButtonContainer.append($deleteButton, $editButton);
-
   return $itemButtonContainer;
 };
